ARM: psci: client: fix power off and reset via 64-bit secure monitor
The PSCI function code is a 32-bit integer > 0x80000000, but the
the PSCI client code erroneously casted it to a signed integer.

On 32-bit systems, the cast later on to unsigned long restored the
original value, but on 64-bit system, we passed a very large 64-bit
number in the secure monitor call because of the sign extension.

This worked because TF-A seems to ignore the upper 32-bits anyway,
(tested with i.MX8MN), but for other secure monitors like the default
HVC implementation for QEMU ARM64 Virt machine this didn't hold true
for all commands, leading to aborts when doing reset or poweroff.

Link: https://lore.barebox.org/20211015160544.5783-1-a.fatoum@pengutronix.de

+++ arch/arm/cpu/psci-client.c
@@ -15,7 +15,7 @@
 
 static struct restart_handler restart;
 
-static void __noreturn psci_invoke_noreturn(int function)
+static void __noreturn psci_invoke_noreturn(ulong function)
 {
 	int ret;
